# frontend/packages/api/openapi/admin.yaml
openapi: 3.1.0
info:
  title: Smart IPTV Admin API
  version: 1.0.0
  description: Same-origin API of the admin SPA. Errors are RFC 9457 problem details.
paths:
...
  /api/v1/admin/categories:
    get:
      operationId: categories_list
      description: "GET /api/v1/admin/categories: for access-profile editors and the
        catalogue."
      summary: List categories, by kind and order
      parameters:
        - in: query
          name: kind
          schema:
            type: string
            enum:
              - live
              - series
              - vod
          description: |-
            * `vod` - Movies
            * `series` - Series
            * `live` - Live TV
        - name: page
          required: false
          in: query
          description: A page number within the paginated result set.
          schema:
            type: integer
        - name: page_size
          required: false
          in: query
          description: Number of results to return per page.
          schema:
            type: integer
        - name: search
          required: false
          in: query
          description: A search term.
          schema:
            type: string
      tags:
        - categories
      security:
        - sessionAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCategoryList"
          description: ""
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Invalid request (VALIDATION_ERROR, or a more specific code).
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not signed in (NOT_AUTHENTICATED).
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not allowed (PERMISSION_DENIED, or a more specific code).
        default:
          description: Error (RFC 9457 problem details).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      operationId: categories_create
      description: "GET /api/v1/admin/categories: for access-profile editors and the
        catalogue."
      summary: Create a category
      tags:
        - categories
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryWriteRequest"
        required: true
      security:
        - sessionAuth: []
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
          description: ""
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Invalid request (VALIDATION_ERROR, or a more specific code).
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not signed in (NOT_AUTHENTICATED).
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not allowed (PERMISSION_DENIED, or a more specific code).
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflicts with the current state (CONFLICT, DEVICE_LIMIT, ...).
        default:
          description: Error (RFC 9457 problem details).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/admin/categories/{id}:
    get:
      operationId: categories_retrieve
      summary: One category
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
      tags:
        - categories
      security:
        - sessionAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
          description: ""
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not signed in (NOT_AUTHENTICATED).
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not allowed (PERMISSION_DENIED, or a more specific code).
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: No such resource (NOT_FOUND).
        default:
          description: Error (RFC 9457 problem details).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      operationId: categories_update
      summary: Change a category
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
      tags:
        - categories
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchedCategoryWriteRequest"
      security:
        - sessionAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
          description: ""
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Invalid request (VALIDATION_ERROR, or a more specific code).
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not signed in (NOT_AUTHENTICATED).
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not allowed (PERMISSION_DENIED, or a more specific code).
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: No such resource (NOT_FOUND).
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflicts with the current state (CONFLICT, DEVICE_LIMIT, ...).
        default:
          description: Error (RFC 9457 problem details).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      operationId: categories_delete
      summary: Delete a category (titles and access profiles lose it)
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
      tags:
        - categories
      security:
        - sessionAuth: []
      responses:
        "204":
          description: No response body
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not signed in (NOT_AUTHENTICATED).
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not allowed (PERMISSION_DENIED, or a more specific code).
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: No such resource (NOT_FOUND).
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflicts with the current state (CONFLICT, DEVICE_LIMIT, ...).
        default:
          description: Error (RFC 9457 problem details).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/admin/categories/reorder:
    post:
      operationId: categories_reorder
      summary: Put the categories of one kind in the given order
      tags:
        - categories
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryReorderRequest"
        required: true
      security:
        - sessionAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
          description: ""
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Invalid request (VALIDATION_ERROR, or a more specific code).
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not signed in (NOT_AUTHENTICATED).
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not allowed (PERMISSION_DENIED, or a more specific code).
        default:
          description: Error (RFC 9457 problem details).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /api/v1/admin/metadata/search:
    get:
      operationId: metadata_search
      summary: Search TMDB (or the fixtures) for manual matching
      parameters:
        - in: query
          name: kind
          schema:
            enum:
              - movie
              - tv
            type: string
            minLength: 1
          description: |-
            * `movie` - Movie
            * `tv` - Series
          required: true
        - in: query
          name: query
          schema:
            type: string
            minLength: 1
            maxLength: 200
          required: true
        - in: query
          name: year
          schema:
            type: integer
            maximum: 2200
            minimum: 1870
      tags:
        - metadata
      security:
        - sessionAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
          description: ""
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Invalid request (VALIDATION_ERROR, or a more specific code).
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not signed in (NOT_AUTHENTICATED).
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not allowed (PERMISSION_DENIED, or a more specific code).
        default:
          description: Error (RFC 9457 problem details).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /api/v1/admin/movies/{id}/refresh-metadata:
    post:
      operationId: movies_refresh_metadata
      summary: Fetch the movie's metadata again (locked fields are kept)
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
      tags:
        - movies
      security:
        - sessionAuth: []
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Queued"
          description: ""
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not signed in (NOT_AUTHENTICATED).
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not allowed (PERMISSION_DENIED, or a more specific code).
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: No such resource (NOT_FOUND).
        default:
          description: Error (RFC 9457 problem details).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /api/v1/admin/review-queue/{id}:
    get:
      operationId: review_queue_retrieve
      summary: One review with the parsed name, the file and the scored candidates
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
      tags:
        - review-queue
      security:
        - sessionAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
          description: ""
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not signed in (NOT_AUTHENTICATED).
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Not allowed (PERMISSION_DENIED, or a more specific code).
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: No such resource (NOT_FOUND).
        default:
          description: Error (RFC 9457 problem details).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
...
    CategoryKind:
      enum:
        - vod
        - series
        - live
      type: string
      description: |-
        * `vod` - Movies
        * `series` - Series
        * `live` - Live TV
    CategoryReorderRequest:
      type: object
      properties:
        kind:
          $ref: "#/components/schemas/CategoryKind"
        ids:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 1000
      required:
        - ids
        - kind
...
    PaginatedCategoryList:
      type: object
      required:
        - count
        - next
        - previous
        - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type:
            - string
            - "null"
          format: uri
        previous:
          type:
            - string
            - "null"
          format: uri
        results:
          type: array
          items:
            $ref: "#/components/schemas/Category"
...
    Queued:
      type: object
      properties:
        queued:
          type: boolean
      required:
        - queued
...
    Problem:
      type: object
      description: RFC 9457 problem details, served as application/problem+json.
      properties:
        type:
          type: string
          description: urn:smart-iptv:problem:<code>
        title:
          type: string
        status:
          type: integer
        code:
          $ref: "#/components/schemas/ErrorCode"
        detail:
          type: string
        field_errors:
          type: object
          description: Messages per field; nested fields use dotted paths.
          additionalProperties:
            type: array
            items:
              type: string
      required:
        - type
        - title
        - status
        - code
        - detail
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: sessionid